Divide get_data() to get raw data
Monitored value is returned as a hex value from a device, so it should
be converted into decimal number with `int(val, 16)`. However, some of
the returned values are not a simple value, but a combination of two
values for example. It means that it's requires to some processes before
converting from hex for these value.

This update adds `get_raw_data()` to get the hex value, and changes
existing `get_data()` to a wrapper method to convert.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>