Fix ampere value to be correct
The value of electric current from the device is returned as a set of
two values in a hex format, T and R phases. For example, "003C0000"
should be divided into "003C" and "0000", then converted to "50" and
"0" for considering the specification of Echonet [1]. Without this
correction, "003C0000" is converted into a wrong value "3932160".

These values will be converted into 5.0[A] and 0.0[A] to become more
understandable finally.

[1] https://echonet.jp/wp/wp-content/uploads/pdf/General/Standard/Release/Release_H_jp/Appendix_H.pdf
    p.312-319

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>