Add InfluxDB as a data store
Although influxdb client provides useful features, add a client class
to do some initial setups for creating DB in which monitored data is
stored.

You should define params for InfluxDB, such as host, username and
password, in `secret/influx.yml'.

  host: YOUR_HOST
  dbname: YOUR_DB
  username: YOUR_ACCOUNT
  password: YOUR_PASSWORD

`host` is used as a tag in registered data, arbitrary value, so you can
use any of values for `host`, for example `localhost` or so.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>